app.py
import os
import gdown
import io
import warnings
import traceback

import numpy as np
from PIL import Image
import uvicorn
from fastapi import FastAPI, File, UploadFile
from fastapi.middleware.cors import CORSMiddleware

import tensorflow as tf
import tf_keras as keras
from tf_keras.applications.mobilenet_v2 import preprocess_input
warnings.filterwarnings('ignore')

app = FastAPI()

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# ============================================
# CUSTOM LOSS FUNCTION
# ============================================
def label_smoothing_loss(y_true, y_pred, smoothing=0.1):
    num_classes = tf.cast(tf.shape(y_true)[-1], tf.float32)
    y_true_smooth = y_true * (1 - smoothing) + (smoothing / num_classes)
    return keras.losses.categorical_crossentropy(y_true_smooth, y_pred)
# ============================================
# DOWNLOAD MODEL FROM GOOGLE DRIVE
# ============================================
def download_model():
    os.makedirs('models', exist_ok=True)
    model_path = 'models/mobilenetv2_model.keras'

    if not os.path.exists(model_path):
        print("⬇️ Downloading model from Google Drive...")
        file_id = "YOUR_GOOGLE_DRIVE_FILE_ID"   # 👈 replace this
        gdown.download(f"https://drive.google.com/drive/folders/1YXrFzrB_ZRM2O1MWNxeOmlZs0ePpxujQ?usp=sharing", model_path, quiet=False)
        print("✅ Model downloaded!")
    else:
        print("✅ Model file already exists, skipping download")

download_model()  # 👈 runs before model is loaded
# ============================================
# LOAD MODELS
# ============================================
print("Loading models...")

model_mobilenet = None
model_hybrid = None

try:
    model_mobilenet = keras.models.load_model(
        'models/mobilenetv2_model.keras',
        custom_objects={'label_smoothing_loss': label_smoothing_loss},
        compile=False
    )
    print("✅ MobileNetV2 loaded")
except Exception as e:
    print(f"❌ Error loading MobileNetV2: {str(e)[:200]}")

print("⚠️ Hybrid model disabled to save memory")

# ============================================
# CLASS LABELS
# ============================================
CLASS_LABELS = [
    'Acne',
    'Eczema',
    'Keratosis',
    'Carcinoma',
    'Milia',
    'Rosacea',
]

NON_SKIN_THRESHOLD = 60.0

# ============================================
# IMAGE PREPROCESSING
# ============================================
def preprocess_image(image_bytes):
    try:
        img = Image.open(io.BytesIO(image_bytes)).convert('RGB')
        img = img.resize((224, 224))
        img_array = np.array(img, dtype=np.float32)
        img_array = preprocess_input(img_array)
        img_array = np.expand_dims(img_array, axis=0)
        return img_array
    except Exception as e:
        print(f"[PREPROCESS ERROR] {e}")
        traceback.print_exc()
        return None

# ============================================
# CONDITION DESCRIPTIONS
# ============================================
def get_condition_description(condition_name):
    descriptions = {
        'Acne': 'Inflammatory skin condition characterized by comedones, papules, pustules, or cysts.',
        'Eczema': 'Atopic dermatitis causing itchy, inflamed, and sometimes scaly patches of skin.',
        'Rosacea': 'Chronic inflammatory condition causing facial redness and visible blood vessels.',
        'Keratosis': 'Rough, scaly patches caused by buildup of keratin.',
        'Carcinoma': 'Abnormal growth that may indicate skin cancer. Requires medical evaluation.',
        'Milia': 'Small, white keratin-filled cysts appearing as tiny bumps.'
    }
    return descriptions.get(condition_name, 'Consult a dermatologist.')

# ============================================
# RECOMMENDATIONS
# ============================================
def get_recommendations(condition_name):
    recommendations = {
        'Acne': [
            'Use non-comedogenic products',
            'Try salicylic acid or benzoyl peroxide',
            'Consult a dermatologist if severe'
        ],
        'Eczema': [
            'Keep skin moisturized',
            'Avoid harsh soaps',
            'Use prescribed creams if needed'
        ],
        'Rosacea': [
            'Avoid triggers',
            'Use gentle skincare products',
            'Consult dermatologist for treatment'
        ],
        'Keratosis': [
            'Use sunscreen daily',
            'Consider retinoid creams',
            'Seek dermatologist advice'
        ],
        'Carcinoma': [
            'URGENT: See a dermatologist immediately',
            'Avoid sun exposure',
            'Do not delay medical evaluation'
        ],
        'Milia': [
            'Avoid squeezing',
            'Use gentle exfoliation',
            'Professional extraction recommended'
        ],
    }
    return recommendations.get(condition_name, ['Consult a dermatologist'])

# ============================================
# GET MODEL PREDICTIONS
# ============================================
def get_model_predictions(model, image_array, model_name, model_type):
    if model is None:
        return None

    try:
        predictions = model.predict(image_array, verbose=0)
        confidences = predictions[0] * 100

        conditions = []
        for i, class_name in enumerate(CLASS_LABELS):
            conditions.append({
                'name': class_name,
                'confidence': float(confidences[i]),
                'description': get_condition_description(class_name),
                'recommendations': get_recommendations(class_name)
            })

        conditions.sort(key=lambda x: x['confidence'], reverse=True)

        metrics = {
            'accuracy': 0.95,
            'precision': 0.93,
            'f1Score': 0.94
        }

        return {
            'modelName': model_name,
            'modelType': model_type,
            'metrics': metrics,
            'predictions': conditions
        }

    except Exception as e:
        print(f"[{model_name}] Prediction error: {e}")
        traceback.print_exc()
        return None

# ============================================
# HEALTH ENDPOINT
# ============================================
@app.get("/health")
async def health():
    return {
        "status": "OK",
        "models_loaded": {
            "mobilenet": model_mobilenet is not None,
            "hybrid": model_hybrid is not None
        }
    }

# ============================================
# MAIN ANALYSIS ENDPOINT
# ============================================
@app.post("/analyze")
async def analyze_skin(image: UploadFile = File(...)):
    try:
        image_bytes = await image.read()
        img_array = preprocess_image(image_bytes)

        if img_array is None:
            return {"error": "Failed to process image"}

        results = []

        if model_mobilenet:
            mobilenet_result = get_model_predictions(
                model_mobilenet, img_array, 'MobileNetV2', 'Baseline Model'
            )
            if mobilenet_result:
                results.append(mobilenet_result)

        if not results:
            return {"error": "All models failed to produce results"}

        return results

    except Exception as e:
        traceback.print_exc()
        return {"error": str(e)}

# ============================================
# RUN SERVER (local dev only)
# ============================================
if __name__ == "__main__":
    port = int(os.environ.get("PORT", 10000))
    uvicorn.run(app, host="0.0.0.0", port=port)