requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow-cpu==2.15.0
tf-keras
pillow
numpy
gdown